+++ Makefile
@@ -4,8 +4,8 @@ CC      = g++
 FLAGS   = -Wall -Wextra -Werror -std=c++98
 SRCDIR  = ./src/
 OBJDIR  = $(SRCDIR)obj/
-SRCS    = server.cpp client.cpp HttpRequest.cpp HttpResponse.cpp webserv.cpp 
-OBJS    = $(addprefix $(OBJDIR), $(SRCS:.cpp=.o))
+SRCS    = server.cpp client.cpp HttpRequest.cpp HttpResponse.cpp webserv.cpp ConfigParser.cpp RequestHandler.cpp
+OBJS    = $(addprefix $(OBJDIR), $(SRCS:.cpp=.o)) 
 RM      = rm -f
 
 .PHONY: all clean fclean re
