+++ src/HttpResponse.cpp
@@ -4,10 +4,34 @@
 HttpResponse::HttpResponse() : _statusCode(200), _statusMessage("OK") {}
 HttpResponse::~HttpResponse() {}
 
-void HttpResponse::setStatusCode(int code) { _statusCode = code; }
 void HttpResponse::setStatusMessage(const std::string& message) { _statusMessage = message; }
 void HttpResponse::addHeader(const std::string& key, const std::string& value) { _headers[key] = value; }
 void HttpResponse::setBody(const std::string& body) { _body = body; }
+
+static std::map<int, std::string> initStatusMessages() {
+    std::map<int, std::string> m;
+    m[200] = "OK";
+    m[201] = "Created";
+    m[204] = "No Content";
+    m[400] = "Bad Request";
+    m[404] = "Not Found";
+    m[405] = "Method Not Allowed";
+    m[413] = "Payload Too Large";
+    m[500] = "Internal Server Error";
+    m[501] = "Not Implemented";
+    m[505] = "HTTP Version Not Supported";
+    return m;
+}
+
+std::map<int, std::string> HttpResponse::_statusMessages = initStatusMessages();
+
+void HttpResponse::setStatusCode(int code) {
+    _statusCode = code;
+    // автоматически подставляем текст
+    std::map<int,std::string>::iterator it = _statusMessages.find(code);
+    _statusMessage = (it != _statusMessages.end()) ? it->second : "Unknown";
+}
+
 
 std::string HttpResponse::buildResponse() const {
     std::stringstream response_ss;
@@ -31,42 +55,42 @@ std::string HttpResponse::buildResponse() const {
     return response_ss.str();
 }
 
-HttpResponse handleHttpRequest(const std::string& raw_request) { // Это надо заменить на RequestHandler.cpp
-    HttpRequest req(raw_request);
-    HttpResponse res;
+//HttpResponse handleHttpRequest(const std::string& raw_request) {
+//    HttpRequest req(raw_request);
+//    HttpResponse res;
 
-    if (!req.isValid()) {
-        res.setStatusCode(400);
-        res.setStatusMessage("Bad Request");
-        res.setBody("Cannot parse request");
-        return res;
-    }
+//    if (!req.isValid()) {
+//        res.setStatusCode(400);
+//        res.setStatusMessage("Bad Request");
+//        res.setBody("Cannot parse request");
+//        return res;
+//    }
 
-    if (req.getMethod() == "GET") {
-        if (req.getUri() == "/") {
-            res.setStatusCode(200);
-            res.setStatusMessage("OK");
-            res.setBody("Welcome to webserv!\n");
-        } else {
-            res.setStatusCode(404);
-            res.setStatusMessage("Not Found");
-            res.setBody("The requested resource was not found.");
-        }
-    } else if (req.getMethod() == "POST") {
-        res.setStatusCode(200);
-        res.setStatusMessage("OK");
-        res.setBody("POST request received.");
-    } else if (req.getMethod() == "DELETE") {
-        res.setStatusCode(200);
-        res.setStatusMessage("OK");
-        res.setBody("DELETE acknowledged.");
-    } else {
-        res.setStatusCode(405);
-        res.setStatusMessage("Method Not Allowed");
-        res.setBody("Unsupported method.");
-    }
+//    if (req.getMethod() == "GET") {
+//        if (req.getUri() == "/") {
+//            res.setStatusCode(200);
+//            res.setStatusMessage("OK");
+//            res.setBody("Welcome to webserv!\n");
+//        } else {
+//            res.setStatusCode(404);
+//            res.setStatusMessage("Not Found");
+//            res.setBody("The requested resource was not found.");
+//        }
+//    } else if (req.getMethod() == "POST") {
+//        res.setStatusCode(200);
+//        res.setStatusMessage("OK");
+//        res.setBody("POST request received.");
+//    } else if (req.getMethod() == "DELETE") {
+//        res.setStatusCode(200);
+//        res.setStatusMessage("OK");
+//        res.setBody("DELETE acknowledged.");
+//    } else {
+//        res.setStatusCode(405);
+//        res.setStatusMessage("Method Not Allowed");
+//        res.setBody("Unsupported method.");
+//    }
 
-    return res;
-}
+//    return res;
+//}
