+++ src/conf
@@ -19,7 +19,7 @@ server {
 
 	location /uploads {
         upload_dir /data/uploads
-		allow_methods GET POST DELET;
+		allow_methods GET POST DELETE;
 		autoindex on;
 		error_page 404 /errors/404.html;
     }
